MOB-1682 Analyse & Cleanup potentials errors
- Remove some useless methods: viewdidunload
- Add super class call in: viewdidappear, will appear, .. in some classes
- Release CF object after utilisation in ShareViewController: method: verification of orientation of image.

+++ Tests/ExoTestCase.m
@@ -41,7 +41,7 @@
 {
     NSMutableString * output = [NSMutableString string];
     const unsigned char * source = (const unsigned char *)[s UTF8String];
-    int sourceLen = strlen((const char *)source);
+    long sourceLen = strlen((const char *)source);
     for (int i = 0; i < sourceLen; ++i) {
         const unsigned char thisChar = source[i];
         if (thisChar == ' '){
